Add files via upload
Abstract: This dissertation seeks to analyse if there is a relationship between cell-sharing and mental wellbeing in the male adult prisoner population in England and Wales. 5981 respondents to HM Inspectorate of Prisons’ 2022/23 Survey represent this population. Quantitative research methods are used in this study to understand this relationship, through statistical analysis and logistic regression of the cross-sectional cohort. The exploration of mental wellbeing within prisons will argue that progression is needed in the prison system, as overcrowding has put a strain on staff and resources, removing the opportunity of rehabilitation for many prisoners. This study finds that prisoners in individual cells are more likely to experience mental health problems. Demographic and contextual factors are also found to affect this relationship. Those found to have the most significant impact on mental wellbeing are age; engagement in meaningful activity (such as education); feeling safe; illicit drug usage; and frequent visits from familial relations. This research reviews each of these factors, and how they may impact someone’s mental wellbeing, depending on their cell status. Theoretical models, such as dramaturgy, are used to explain that prisoners in shared cells have no opportunity to revert to their backstage self.